AI-Personal-Email-Assistant/requirements.txt
openai
auth
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
slack-sdk
gradio
flask
python-dotenv
requests
